Project files/requirements.txt.txt
tensorflow==2.13.0
keras==2.13.1
matplotlib
numpy
pandas
scikit-learn
opencv-python
Pillow
seaborn
jupyterlab
notebook
flask
gunicorn
streamlit
tensorflow-hub
tensorflow-datasets